No need for HA. Move migrations outside
