+++ ledger/participant-integration-api/src/main/scala/platform/indexer/parallel/InitializeParallelIngestion.scala
@@ -4,11 +4,13 @@
 package com.daml.platform.indexer.parallel
 
 import akka.NotUsed
-import akka.stream.scaladsl.Source
+import akka.stream.Materializer
+import akka.stream.scaladsl.{Sink, Source}
+import com.daml.ledger.api.domain
 import com.daml.ledger.offset.Offset
 import com.daml.ledger.participant.state.v2.{ReadService, Update}
 import com.daml.lf.data.Ref
-import com.daml.logging.LoggingContext
+import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.metrics.Metrics
 import com.daml.platform.ParticipantInMemoryState
 import com.daml.platform.store.backend.{
@@ -28,14 +30,30 @@ private[platform] case class InitializeParallelIngestion(
     metrics: Metrics,
 ) {
 
+  private val logger = ContextualizedLogger.get(classOf[InitializeParallelIngestion])
+
   def apply(
       dbDispatcher: DbDispatcher,
       participantInMemoryState: ParticipantInMemoryState,
       readService: ReadService,
       ec: ExecutionContext,
+      mat: Materializer,
   )(implicit loggingContext: LoggingContext): Future[InitializeParallelIngestion.Initialized] = {
     implicit val executionContext: ExecutionContext = ec
     for {
+      initialConditions <- readService.ledgerInitialConditions().runWith(Sink.head)(mat)
+      providedLedgerId = domain.LedgerId(initialConditions.ledgerId)
+      _ = logger.info(
+        s"Attempting to initialize with ledger ID $providedLedgerId and participant ID $providedParticipantId"
+      )
+      _ <- dbDispatcher.executeSql(metrics.daml.index.db.initializeLedgerParameters)(
+        parameterStorageBackend.initializeParameters(
+          ParameterStorageBackend.IdentityParams(
+            ledgerId = providedLedgerId,
+            participantId = domain.ParticipantId(providedParticipantId),
+          )
+        )
+      )
       ledgerEnd <- dbDispatcher.executeSql(metrics.daml.index.db.getLedgerEnd)(
         parameterStorageBackend.ledgerEnd
       )
