+++ ledger/participant-integration-api/src/main/scala/platform/indexer/parallel/ParallelIndexerFactory.scala
@@ -109,6 +109,7 @@ object ParallelIndexerFactory {
             dbDispatcher = dbDispatcher,
             participantInMemoryState = participantInMemoryState,
             readService = readService,
+            mat = mat,
             ec = ec,
           ).map(
             parallelIndexerSubscription(
